AnnotationPlatform/backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
uvloop==0.21.0
httptools==0.6.4
python-multipart==0.0.19
websockets==14.1

# ── Database ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
greenlet==3.1.1

# ── Auth ───────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# ── Validation & Serialization ─────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# ── Object Storage ─────────────────────────────────────────
minio==7.2.12

# ── Vector Database ────────────────────────────────────────
qdrant-client==1.12.1
grpcio==1.68.1

# ── Redis / Cache ──────────────────────────────────────────
redis[hiredis]==5.2.1

# ── Task Queue ─────────────────────────────────────────────
celery[redis]==5.4.0

# ── ML / Embeddings ────────────────────────────────────────
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
sentence-transformers==3.3.1
open-clip-torch==2.29.0
Pillow==11.0.0
numpy==2.1.3

# ── Utilities ──────────────────────────────────────────────
aiohttp==3.11.11
beautifulsoup4==4.12.3
python-magic==0.4.27
structlog==24.4.0
tenacity==9.0.0
orjson==3.10.12
